% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fairness}
\alias{fairness}
\title{Perveived Fairness of rules for allocating seats in trains and parking spaces}
\format{
a tibble containing:
\itemize{
\item id: the individual index
\item answer: a factor with levels 0 (very unfair), 1 (essentially unfair), 2 (essentially fair) and 3 (very fair)
\item good: one of \code{'tgv'} (French fast train) and \code{'Parking'}
\item rule: the allocation rule, a factor with levels \code{'peak'}, \code{'admin'}, \code{'lottery'}, \code{'addsupply'}, \code{'queuing'}, \code{'moral'} and \code{'compensation'}
\item driving: does the individual has the driving license ?
\item education: does the individual has a diploma ?
\item recurring: does the allocation problem is reccuring ?
}
}
\source{
provided by the authors
}
\description{
a pseudo-panel of 401 individuals from 2003
}
\references{
\insertRef{RAUX:SOUC:CROIS:09}{pglm}
}
\keyword{dataset}
