% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{health_ins}
\alias{health_ins}
\title{Health Insurance and Doctor Visits}
\format{
a tibble containing:
\itemize{
\item id: the individual index
\item year: the year
\item mdu: number of outpatient visits to an MD
\item opu: number of outpation visits to all providers
\item coins: coinsurance rate (0, 25, 50 or 100 percent)
\item idp: if individual deductible plan: 1, otherwise 0
\item lpi: log of the max of 1 and annual participation incentive payment
\item fmde: if idp = 1: 0 otherwise ln of the max of 1 and MDE / (0.01 coins)
\item income: family income
\item size: family size
\item age: the age
\item sex: a factor with level \code{'male'} and \code{'female'}
\item child: a factor with levels \code{'no'} and \code{'yes'}
\item race: a factor with levels \code{'white'} and \code{'black'}
\item health: self-rated health, a factor with levels \code{poor}, \code{fair}, \code{good} and \code{verygood}
\item educ: education of the household head in years
\item physlim: if the person has a physical limitation: 1
\item disease: index of chronic diseases
}
}
\source{
\url{http://cameron.econ.ucdavis.edu/musbook/mus.html}
}
\description{
a cross-section of 5908 individuals from 1974 to 1982
}
\references{
\insertRef{MANN:NEWH:KEEL:LEIB:87}{pglm}

\insertRef{DEB:TRIV:02}{pglm}
}
\keyword{dataset}
