% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hedonic}
\alias{hedonic}
\title{Hedonic Prices of Census Tracts in the Boston Area}
\format{
a tibble containing:
\itemize{
\item mv: median value of owner--occupied homes
\item crim: crime rate
\item zn: proportion of 25,000 square feet residential lots
\item indus: proportion of no--retail business acres
\item chas: is the tract bounds the Charles River?
\item nox: annual average nitrogen oxide concentration in parts per hundred million
\item rm: average number of rooms
\item age: proportion of owner units built prior to 1940
\item dis: weighted distances to five employment centers in the Boston area
\item rad: index of accessibility to radial highways
\item tax: full value property tax rate ($/$10,000)
\item ptratio: pupil/teacher ratio
\item blacks: proportion of blacks in the population
\item lstat: proportion of population that is lower status
\item townid: town identifier
}
}
\source{
Online complements to Baltagi (2013): \url{https://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}
\description{
a cross-section of 506 census tracts
}
\references{
\insertRef{BALT:01}{pglm}

\insertRef{BALT:13}{pglm}

\insertRef{BESL:KUH:WELS:80}{pglm}

\insertRef{HARR:RUBI:78}{pglm}
}
\keyword{dataset}
