% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{patents_rd}
\alias{patents_rd}
\title{Dynamic Relation Between Patents and R&D}
\format{
a tibble containing:
\itemize{
\item cusip: Compustat's identifying number for the firm
\item year: year
\item ardssic: a two-digit code for the applied R
\item scisect: is the firm in the scientific sector ?
\item capital72: book value of capital in 1972
\item sumpat: the sum of patents applied for between 1972-1979
\item rd: R and D spending during the year (in 1972 dollars)
\item patents: the number of patents applied for during the year that were eventually granted
}
}
\source{
\url{http://cameron.econ.ucdavis.edu/racd/racddata.html}, chapter 9.
}
\description{
yearly observations of 346 production units
}
\references{
\insertRef{HALL:GRIL:HAUS:86}{pglm}
}
\keyword{dataset}
