% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{union_wage}
\alias{union_wage}
\title{Unionism and wage rate determination}
\format{
a tibble containing:
\itemize{
\item id: the individual index
\item year: the year
\item exper: the experience, computed as age - 6 - schooling
\item health: does the individual has health disability ?
\item hours: the number of hours worked
\item married: is the individual married ?
\item rural: does the individual lives in a rural area ?
\item school: years of schooling
\item union: does the wage is set by collective bargaining
\item wage: hourly wage in US dollars
\item sector: one of agricultural, mining, construction, trade, transportation, finance,  businessrepair, personalservice, entertainment, manufacturing, pro.rel.service, pub.admin
\item occ: one of proftech, manoffpro, sales, clerical, craftfor, operative, laborfarm, farmlabor, service
\item com: one of black, hisp and other
\item region: the region, one of NorthEast, NothernCentral, South and other
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
yearly observations of 545 individuals from 1980 to 1987
}
\references{
\insertRef{VELL:VERB:98}{pglm}
}
\keyword{dataset}
